// src/gateway/oegw-api.d.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface UpdateThingsResponse {
  /**
   * Message type.
   */
  mType: "updateThings";
  data: {
    things: {
      /**
       * Identificator of thing...
       */
      name: string;
      /**
       * Value of thing...
       */
      value: number;
      /**
       * Type of thing...
       */
      atype: string;
      /**
       * Unit of thing value...
       */
      unit?: string;
      [k: string]: any;
    }[];
    [k: string]: any;
  };
  [k: string]: any;
}